requirements.txt
jax
jaxlib
flax
optax
qiskit
qiskit[visualization]
matplotlib
seaborn
wandb
python-dotenv
pyyaml
pytest
